backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
PyJWT==2.9.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
